plantypes packaged moved to plan.types
