+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/TornadoExecutionPlan.java
@@ -23,27 +23,27 @@ import uk.ac.manchester.tornado.api.common.TornadoDevice;
 import uk.ac.manchester.tornado.api.enums.ProfilerMode;
 import uk.ac.manchester.tornado.api.enums.TornadoVMBackendType;
 import uk.ac.manchester.tornado.api.exceptions.TornadoExecutionPlanException;
-import uk.ac.manchester.tornado.api.plantype.ExecutionPlanType;
-import uk.ac.manchester.tornado.api.plantype.OffConcurrentDevices;
-import uk.ac.manchester.tornado.api.plantype.OffMemoryLimit;
-import uk.ac.manchester.tornado.api.plantype.OffPrintKernel;
-import uk.ac.manchester.tornado.api.plantype.OffProfiler;
-import uk.ac.manchester.tornado.api.plantype.OffThreadInfo;
-import uk.ac.manchester.tornado.api.plantype.WithWarmUp;
-import uk.ac.manchester.tornado.api.plantype.WithBatch;
-import uk.ac.manchester.tornado.api.plantype.WithClearProfiles;
-import uk.ac.manchester.tornado.api.plantype.WithCompilerFlags;
-import uk.ac.manchester.tornado.api.plantype.WithConcurrentDevices;
-import uk.ac.manchester.tornado.api.plantype.WithDefaultScheduler;
-import uk.ac.manchester.tornado.api.plantype.WithDevicePlan;
-import uk.ac.manchester.tornado.api.plantype.WithDynamicReconfiguration;
-import uk.ac.manchester.tornado.api.plantype.WithFreeDeviceMemory;
-import uk.ac.manchester.tornado.api.plantype.WithGridScheduler;
-import uk.ac.manchester.tornado.api.plantype.WithMemoryLimit;
-import uk.ac.manchester.tornado.api.plantype.WithPrintKernel;
-import uk.ac.manchester.tornado.api.plantype.WithProfiler;
-import uk.ac.manchester.tornado.api.plantype.WithResetDevice;
-import uk.ac.manchester.tornado.api.plantype.WithThreadInfo;
+import uk.ac.manchester.tornado.api.plan.types.ExecutionPlanType;
+import uk.ac.manchester.tornado.api.plan.types.OffConcurrentDevices;
+import uk.ac.manchester.tornado.api.plan.types.OffMemoryLimit;
+import uk.ac.manchester.tornado.api.plan.types.OffPrintKernel;
+import uk.ac.manchester.tornado.api.plan.types.OffProfiler;
+import uk.ac.manchester.tornado.api.plan.types.OffThreadInfo;
+import uk.ac.manchester.tornado.api.plan.types.WithWarmUp;
+import uk.ac.manchester.tornado.api.plan.types.WithBatch;
+import uk.ac.manchester.tornado.api.plan.types.WithClearProfiles;
+import uk.ac.manchester.tornado.api.plan.types.WithCompilerFlags;
+import uk.ac.manchester.tornado.api.plan.types.WithConcurrentDevices;
+import uk.ac.manchester.tornado.api.plan.types.WithDefaultScheduler;
+import uk.ac.manchester.tornado.api.plan.types.WithDevicePlan;
+import uk.ac.manchester.tornado.api.plan.types.WithDynamicReconfiguration;
+import uk.ac.manchester.tornado.api.plan.types.WithFreeDeviceMemory;
+import uk.ac.manchester.tornado.api.plan.types.WithGridScheduler;
+import uk.ac.manchester.tornado.api.plan.types.WithMemoryLimit;
+import uk.ac.manchester.tornado.api.plan.types.WithPrintKernel;
+import uk.ac.manchester.tornado.api.plan.types.WithProfiler;
+import uk.ac.manchester.tornado.api.plan.types.WithResetDevice;
+import uk.ac.manchester.tornado.api.plan.types.WithThreadInfo;
 import uk.ac.manchester.tornado.api.runtime.ExecutorFrame;
 import uk.ac.manchester.tornado.api.runtime.TornadoRuntimeProvider;
