+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/OffMemoryLimit.java
@@ -15,19 +15,19 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class WithConcurrentDevices extends ExecutionPlanType {
+public final class OffMemoryLimit extends ExecutionPlanType {
 
-    public WithConcurrentDevices(TornadoExecutionPlan parent) {
+    public OffMemoryLimit(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withConcurrentDevices ";
+        return parentLink.toString() + "\n -> withoutMemoryLimit ";
     }
 
 }
