+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/OffPrintKernel.java
@@ -15,17 +15,17 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class WithCompilerFlags extends ExecutionPlanType {
-    public WithCompilerFlags(TornadoExecutionPlan parent) {
+public final class OffPrintKernel extends ExecutionPlanType {
+    public OffPrintKernel(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withCompilerFlags ";
+        return parentLink.toString() + "\n -> withoutPrintKernel ";
     }
 }
