+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithClearProfiles.java
@@ -15,19 +15,17 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class OffConcurrentDevices extends ExecutionPlanType {
-
-    public OffConcurrentDevices(TornadoExecutionPlan parent) {
+public final class WithClearProfiles extends ExecutionPlanType {
+    public WithClearProfiles(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withoutConcurrentDevices ";
+        return parentLink.toString() + "\n -> withClearProfiles ";
     }
-
 }
