+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithConcurrentDevices.java
@@ -15,17 +15,19 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class WithThreadInfo extends ExecutionPlanType {
-    public WithThreadInfo(TornadoExecutionPlan parent) {
+public final class WithConcurrentDevices extends ExecutionPlanType {
+
+    public WithConcurrentDevices(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withThreadInfo ";
+        return parentLink.toString() + "\n -> withConcurrentDevices ";
     }
+
 }
