+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithDefaultScheduler.java
@@ -15,18 +15,18 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class WithDevicePlan extends ExecutionPlanType {
+public final class WithDefaultScheduler extends ExecutionPlanType {
 
-    public WithDevicePlan(TornadoExecutionPlan parent) {
+    public WithDefaultScheduler(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withDevice ";
+        return parentLink.toString() + "\n -> withDefaultGridScheduler ";
     }
 }
