+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithFreeDeviceMemory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
