+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithPrintKernel.java
@@ -15,17 +15,17 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class WithClearProfiles extends ExecutionPlanType {
-    public WithClearProfiles(TornadoExecutionPlan parent) {
+public final class WithPrintKernel extends ExecutionPlanType {
+    public WithPrintKernel(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withClearProfiles ";
+        return parentLink.toString() + "\n -> withPrintKernel ";
     }
 }
