+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithProfiler.java
@@ -15,19 +15,18 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class OffProfiler extends ExecutionPlanType {
+public final class WithProfiler extends ExecutionPlanType {
 
-    public OffProfiler(TornadoExecutionPlan parent) {
+    public WithProfiler(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withoutProfiler ";
+        return parentLink.toString() + "\n -> withProfiler ";
     }
-
 }
