+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithThreadInfo.java
@@ -15,17 +15,17 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class OffPrintKernel extends ExecutionPlanType {
-    public OffPrintKernel(TornadoExecutionPlan parent) {
+public final class WithThreadInfo extends ExecutionPlanType {
+    public WithThreadInfo(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withoutPrintKernel ";
+        return parentLink.toString() + "\n -> withThreadInfo ";
     }
 }
