+++ tornado-api/src/main/java/uk/ac/manchester/tornado/api/plan/types/WithWarmUp.java
@@ -15,17 +15,18 @@
  * limitations under the License.
  *
  */
-package uk.ac.manchester.tornado.api.plantype;
+package uk.ac.manchester.tornado.api.plan.types;
 
 import uk.ac.manchester.tornado.api.TornadoExecutionPlan;
 
-public final class OffThreadInfo extends ExecutionPlanType {
-    public OffThreadInfo(TornadoExecutionPlan parent) {
+public final class WithWarmUp extends ExecutionPlanType {
+
+    public WithWarmUp(TornadoExecutionPlan parent) {
         super(parent);
     }
 
     @Override
     public String toString() {
-        return parentLink.toString() + "\n -> withoutThreadInfo ";
+        return super.toString() + "\n -> withWarmUp ";
     }
 }
