+++ tornado-unittests/src/main/java/uk/ac/manchester/tornado/unittests/executor/TestExecutor.java
@@ -286,7 +286,6 @@ public class TestExecutor extends TornadoTestBase {
             String trace1 = trace.getTraceExecutionPlan();
             String trace2 = executionPlan.getTraceExecutionPlan();
             assertEquals(trace1, trace2);
-
         }
     }
